+++ apps/mobile/src/screens/measurement-flow-screen/components/flow-nodes/analysis-node/analysis-node.tsx
